- LuckPerm Events are now listened to, to make sure they are given the correct Context - PermissionImplLP is now on the server thread - Optimized and improved PermissionImplLP
New Config Option: Context
Context is used for catching permissions which are given to a Player without the Context. Context is used to define which Server a Player has a permission on, so permissions don't travel to each Gamemode.

Reorganized everything because the Plugin is being upgraded to cover the entire Network